pyproject.toml
[project]
name = "itxda-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.3.2",
    "pandas>=2.3.1",
    "prefect-client>=3.4.12",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.9",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.44",
    "sshtunnel>=0.4.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pandas-stubs>=2.3.0.250703",
    "types-pymysql>=1.1.0.20250916",
]
